requirements.txt
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
xgboost==2.0.3
lightgbm==4.1.0
scikit-learn==1.3.2
vaderSentiment==3.3.2
textblob==0.17.1
tvdatafeed==2.1.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
ta==0.10.2
yfinance==0.2.28
beautifulsoup4==4.12.2
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2
pyyaml==6.0.1